requirements.txt
streamlit==0.80.0
Pillow==8.3.1
